requirements.txt
python-telegram-bot==20.2
libtorrent==2.0.7
torrentp==0.1.5
requests==2.28.2
gspread == 5.8.0
oauth2client==4.1.3
